+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## Unreleased
 
 - First release
+- ZMQ distributed queue